Rests and durations.
Can now render rests, although I started out rendering it simply as the
placeholder text "rest". It's nice to actually see it in place, but will
eventually be replaced by the actual icons.

Durations are now a proper class. Eventually they'll support operations to
mutate them (dotting, n-tuplet, and so on).

Also removed the divisions attribute, since it's not needed for rendering. I
just took it from MusicXML for simplicity, but it was meant for sound output.