Load and draw dead notes.
Also, we load and draw linked notes (ties), with a period as the placeholder
text. Doing this for now since rendering ties is a little challenging at the
moment. In particular, a linked note needs to know the starting note before we
can draw the tie. This will also have to be an "overlay" over the staff group,
which is a concept we currently don't have.