Render ghost notes.
Also moved stringification of a note to the `Note` class. This allows computing
a width in layout based on the textual size of the note. I previously
implemented a slightly more efficient computation of the width, but we can
return to optimizing performance here when performance is an actual concern.