Add two new layouts to simplify future additions.
The first group is a stacking group. It's kind of like a flex group, but its own
box grows with the elements added, instead of being a preset box with extra
space divided among the children. This is useful for the elements above a staff,
which are often stacked (lyrics, measure number, tempo changes, and so on).

The second layout allows anchoring the location of the elements in the layout to
some other element. This is useful for elements above/below the staff, which
need to be aligned with a corresponding chord/note.