Refactor line and grid layouts to support non-chord elements.
In particular, we want there to always be a left edge at the beginning of the
line, but the vertical "TAB" text means the first chord isn't positioned close
to 0. This is solved by injecting a grid layout element (in `Line`) that will
map to 0, but requires updating a lot of other code relying on
`gridLayoutElements` to check for measure/element nil-ness.

A couple of other changes crept into this one:

1. Created an abstract group class that does no layout, but acts as a basis from
   which many other layouts can derive from.
2. Groups now own the boxes they are given. That means the caller has to no
   longer need ownership, or clone if they want to avoid having the box mutated.
   This surfaced one problematic area where arcs shared the same box as the
   staff.