Improve decorations that may cover more than a single chord.
In particular, for the element generator function, we pass an additional boolean
to indicate whether or not the predicate applied to more than a single value.
For things that use a dashed line to indicate it covers more than one chord, we
only want to use the dashed line element if more than one chord passes the
predicate.