Fix playback box.
Since we moved to queuing up all notes for a measure, we lost the playback box
highlighting the currently played chord. There's no straightforward way to do
this right now, at least not one that would sync with the audio context timer,
so we instead just show the measure currently being played instead.