Fix bug in `Box#encompass`.
If no boxes given, we want to default to an empty box, not an undefined one.

If only one box given, we don't want that box instance to be returned, but a new
one. This was causing issues when only one page was visible in a part, which
made the part's box be the page's box, and we want to mutate the former.