Break up page line into multiple components.
One thing I don't like about this is that I have to pass information from the
`PageLine` component to these children element. That means there's some level of
coupling, but I'm okay with that because these form a package, so as long as the
coupling is within that package, s'all good.

One nice thing from this refactor was realizing `AnchoredGroup` wasn't actually
being used, so we could delete a bunch of code. Woo!