Remove unnecessary function `BelowStaff#elementOffset`
When we moved to having chords accompanied by a growing spacer element, to give
a visual indication of the duration of the chord, the box of the chord was now
only as wide as the notes. This means we can always just use the center of the
beat element for stems, beaming, and dotting.