Apply coarse tune generator to pitches.
This generator is meant as a pitch offset, and was pronounced in a local sound
font I had where the bass had this generator but I wasn't applying it. The
result was that most notes ended up sounding the exact same, which is perhaps
another bug to address (don't play anything if a note can't be found).