Fix rendering bug for instruments not having 6 strings.
We assumed a chord height for instruments with 6 strings, but guitar tabs
commonly come with a bass track too. For now, we simply make the chord height
max out based on the note's placement value. Ideally, we can access the
instrument for the chord and use the number of strings from it, because:

1. all chords are the same height, but more importantly,
2. we'd have a box to check for clicks within a chord.