Refactor routing logic and components.
First, we lifted out the router from `<Score>`. This meant factoring out the
initial page, which was made easier by having a shared SVG root component.

Second, we switched to hash-based routing. Muzart is just a single-page app at
the moment, so we don't want the underlying server think we're trying to go to
some actual underlying route (that will never exist).