Try to bring some sense to the wheel/scroll events.
The zoom effect we have can happen if the user swipes hard on a track pad and
then later presses the meta key. This is because the initial track pad swipe
produces many scroll/wheel events, but we need to detect what the first one is
and whether the meta key was pressed on that event.

On top of this, there's no concept of a start/end, so what we do is time out
after the initial event to say "okay, all events are now done for the initial
action that started them". Also, it's hard to know when font zooming will occur
versus scrolling, so we prevent all defaults when there's a vertical delta and
manually scroll.

It's not perfect, but it gets the job done.