Don't use `useAsync` for loading scores.
The problem is that if the hook fires for the same URL twice, the first promise
that is actually doing the load is ignored, and we don't get a proper result for
the second one (technically, it resolves immediately without a value).

There could be ways to solve this with `useAsync`, but we've already built nice
async state things into `Application`, so why not use them?