Simplify Karplus-Strong configuration
There's a bunch of other, sorta unrelated changes in here too.

This commit kind of breaks things. I took away the envelopes in favour of a
proper duration. This was because the envelopes were kind of redundant:

* Sampler instruments have an envelope baked into the soundfont.
* Karplus-Strong (KS) naturally fades over time.
* Drums built on KS also abruptly stop.

Another problem I've observed is that the duration of tied notes is wrong. We
simply sum up the note values, ignoring their positions within the measure and
all the notes in between (necessary to clearly express the other notes + beat).