Make resizing the canvas less janky.
Previously, when the canvas was resized it would stretch until the user stopped
resizing. After 100ms passed, it would settle into the new state, possibly
jumping around some.

To avoid this jankiness, we make a canvas as large as it ever could be (the
screen dimensions) and use overflow to clip it. We compute our viewport based on
the visible portion, but now when we resize we don't have to change the
width/height attributes of the canvas, so we avoid stretchy artifacts and such.