Improve resize handling of canvas.
We've removed a resize observer that attempts to adjust the viewport to adapt to
changes in the canvas's parent element. It unfortunately wasn't doing a great
job and often didn't maintain the desired aspect ratio.

Instead, we install a resize observer in the canvas state when we set a new
canvas, and this will keep track of the parent element's width/height. This
allows us to adjust the aspect ratio as necessary when the parent element
changes.

One other added benefit is that the `canvasWidth` and `canvasHeight` properties
are now observable!